Added viewport for listing washes
The viewport area is the main focus of the web application.
It will display a list of washes, a form for creating washes,
list tools, a form for creating tools, etc..

It will be used to display the primary action the user has triggered.

This commit implements a viewport in App.js, which can be used to
display content implemented by other components.